WA (same): Changed to double but the error still remains

+++ ADTM202405302030/d.cpp
@@ -10,15 +10,15 @@ int main() {
 
 	vector<vector<int>> rotation_mtx = vector<vector<int>>(2, vector<int>(2, 0));
 
-	float d_radian = d / (float)180 * M_PI;
+	double d_radian = d / (double)180 * M_PI;
 
 	rotation_mtx[0][0] = cos(d_radian);
 	rotation_mtx[0][1] = -sin(d_radian);
 	rotation_mtx[1][0] = sin(d_radian);
 	rotation_mtx[1][1] = cos(d_radian);
 
-	float a_dash = cos(d_radian) * (float)a - sin(d_radian) * (float)b;
-	float b_dash = sin(d_radian) * (float)a + cos(d_radian) * (float)b;
+	double a_dash = cos(d_radian) * (double)a - sin(d_radian) * (double)b;
+	double b_dash = sin(d_radian) * (double)a + cos(d_radian) * (double)b;
 
 	cout << a_dash << " " << b_dash << endl;
